fix: fake native tokens need better UX error handling (#8572)
## **Description**

In response to recent scam incidents we're adding mitigations to help
prevent users from adding malicious networks and being shown inaccurate
token values.

Implementation of this ticket will:

Show N/A for a native token value, when the native token symbol does not
match the chain ID as publicly known
Show a tooltip that when tapped shows a bottom sheet explainer
The explainer directs the user to update their Network details for the
network the token balance is requested for
User story

As a user I want to be informed that a token balance might be
inaccurate, so that I don't make decisions to transfer or buy assets
based on this information
As a user I want to be informed when a custom network might be trying to
scam me


## **Related issues**

Fixes: [#1354](https://github.com/MetaMask/mobile-planning/issues/1354)

## **Manual testing steps**

1. Go to edit network page and choose Polygon
2.change the symbol from MATIC to ETH
3. Go to the wallet page

## **Screenshots/Recordings**

<!-- If applicable, add screenshots and/or recordings to visualize the
before and after of your change. -->

### **Before**


https://github.com/MetaMask/metamask-mobile/assets/26223211/af65f12b-0430-443a-9631-7b8beb188ee8


### **After**


https://drive.google.com/drive/folders/1lcKOyC7U1r9OW2qt7BjsReGdG-02EysJ?usp=sharing


<!-- [screenshots/recordings] -->

## **Pre-merge author checklist**

- [x] I’ve followed [MetaMask Coding
Standards](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've clearly explained what problem this PR is solving and how it
is solved.
- [x] I've linked related issues
- [x] I've included manual testing steps
- [x] I've included screenshots/recordings if applicable
- [x] I’ve included tests if applicable
- [x] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [x] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.
- [x] I’ve properly set the pull request status:
  - [ ] In case it's not yet "ready for review", I've set it to "draft".
- [x] In case it's "ready for review", I've changed it from "draft" to
"non-draft".

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.