fix(8667): trigger swap tx on approval tx confirmed (#9173)
## **Description**

When swapping a token that needs approval (and requires a 2 step swap
process), it's needed for the user, when using a hardware wallet, to
manually sign transactions. Currently, second modal opening is triggered
when approval transaction is finished (meaning when it has been
submitted). But this would end up in a situation where the first modal
was not unmounted before it was asked to be mounted again. This flow was
creating an issue where the user could not confirm the second
transaction (as second modal was never displayed)

To mitigate this issue, we choose to wait for first transaction
completion before starting the second one. This is not perfect from user
POV as a long delay can happen between approval and swap txs, but as
discussed with @AlexJupiter and @gantunesr it's considered as a
temporary solution until we figure out a long term one.

## **Related issues**

Fixes: https://github.com/MetaMask/metamask-mobile/issues/8667

## **Manual testing steps**
For ease the explanation, we are using `AAVE` and `MATIC` on Polygon
network and will sign txs with a Ledger account. You also need blind
signing to be enabled on your Ledger device.

1. make sure `AAVE` token approval has been revoked here
https://polygonscan.com/tokenapprovalchecker for your Ledger account
2. select swap action from actions dropdown
3. select your Ledger account
4. select `AAVE` token as origin token
5. type a value that you want to swap (`0.0000001` is fine)
6. select `MATIC` token as destination token
7. get quotes (you may need to retry several times this step)
8. Swipe to swap
9. First modal should be displayed. Sign approval tx on your ledger
device
10. Once approval txs is confirmed, second modal is opening. Sign swap
tx on your ledger device
11. Swap should be completed

## **Screenshots/Recordings**

### **Before**
https://recordit.co/Vg7MLtediW

### **After**

https://github.com/MetaMask/metamask-mobile/assets/574787/fbe01613-2dd5-4762-98f6-4072782e7491


## **Pre-merge author checklist**

- [x] I’ve followed [MetaMask Coding
Standards](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've completed the PR template to the best of my ability
- [ ] I’ve included tests if applicable
- [ ] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [x] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.